Bugs almost fixed in usersController

+++ app/Http/Controllers/v1/usersController.php
@@ -115,31 +115,38 @@ class usersController extends Controller
         $validator = Validator::make($request->all(),[
             'name' => 'required|',
             'email' => 'required|email|unique:users',
-            'user_type' => 'required',
+
         ]);
 
         if ($validator->fails()){
             return response()->json([
-                'error' => $validator->errors()
+                'error' => $validator->errors(),
+                $request->request,
             ], 401);
         }
 
-        try {
 
-            $user->update([
-                'name' => $request->name,
-                'email' => $request->email,
-                'user_type' => $request->user_type
-            ]);
-            return response()->json([
-                'message' => 'user '.$user->name.' updated'
+            if(Auth::id()!= $user->id){
+                return response()->json([
+                    'message' => 'user unauthorized',
+                ], 401);
+            }else {
+                try {
 
-            ],200);
-        } catch (Exception $e){
-            return response()->json([
-                'error' => $e
-            ],500);
-        }
+                    $user->update([
+                        'name' => $request->name,
+                        'email' => $request->email,
+                    ]);
+                    return response()->json([
+                        'message' => 'user ' . $user->name . ' updated'
+
+                    ], 200);
+                } catch (Exception $e) {
+                    return response()->json([
+                        'error' => $e
+                    ], 500);
+                }
+            }
     }
 
     /**
@@ -150,11 +157,18 @@ class usersController extends Controller
      */
     public function destroy(User $user)
     {
-        $user->delete();
-        return response()->json([
-            'message' => 'Deletion Success',
-            'duration' => '7',
-        ],200);
+        if(Auth::id()!=$user->id){
+            return response()->json([
+                'message' => 'not authorized',
+            ],401);
+        }else{
+            $user->delete();
+            return response()->json([
+                'message' => 'Deletion Success',
+                'duration' => '7',
+            ],200);
+        }
+
     }
 
 //    public function updateState()
